Bug Fix: Refining equipment option didn't not accept decimal values (Issue #340)

+++ src/main/java/net/nikr/eve/jeveasset/io/local/SettingsReader.java
@@ -1143,7 +1143,7 @@ public final class SettingsReader extends AbstractXmlReader<Boolean> {
 				ore = 0;
 			}
 		}
-		int station = getInt(windowElement, "station");
+		double station = getDouble(windowElement, "station");
 		settings.setReprocessSettings(new ReprocessSettings(station, reprocessing, efficiency, ore, scrapmetal));
 	}
